Simplified macros and enums (I think)...

+++ src/item_menu.c
@@ -916,7 +916,7 @@ static void GetItemNameFromPocket(u8 *dest, u16 itemId)
     case POCKET_TM_HM:
         end = StringCopy(gStringVar2, GetMoveName(ItemIdToBattleMoveId(itemId)));
         PrependFontIdToFit(gStringVar2, end, FONT_NARROW, 61);
-        if (gItemsInfo[itemId].tmHmIndex >= (u8)NUM_TECHNICAL_MACHINES)
+        if (gItemsInfo[itemId].tmHmIndex >= NUM_TECHNICAL_MACHINES)
         {
             // Get HM number
             ConvertIntToDecimalStringN(gStringVar1, gItemsInfo[itemId].tmHmIndex - NUM_TECHNICAL_MACHINES + 1, STR_CONV_MODE_LEADING_ZEROS, 1);
